Move from LibCSo6 to LinuxHelper
